requirements.txt
tensorflow==2.9.1
dvc
pandas
notebook
numpy
matplotlib
seaborn
python-box
pyYAML
tqdm
ensure
joblib
types-PyYAML
scipy>=1.9.1
mlflow==1.26.1

-e .